pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "sentinel-oversight-openenv"
version = "1.0.0"
description = "OpenEnv environment for production incident response and SENTINEL AI oversight training."
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.11"
keywords = ["openenv", "reinforcement-learning", "incident-response", "sre", "rl-environment"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.34.0",
    "pydantic>=2.10.0",
    "httpx>=0.28.0",
    "openai>=2.7.2",
    "python-dotenv>=1.0.0",
    "pyyaml>=6.0.0",
    "openenv-core>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
]
train = [
    "torch>=2.0.0",
    "transformers>=4.45.0",
    "peft>=0.13.0",
    "trl>=0.12.0",
    "datasets>=2.20.0",
    "matplotlib>=3.8.0",
    "wandb>=0.17.0",
]

[project.scripts]
server = "server.app:main"

[project.urls]
Homepage = "https://huggingface.co/spaces/srikrishna2005/openenv"
Repository = "https://github.com/sri11223/openEnv"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "baseline*", "server*", "sentinel*", "training*", "judges*"]

[tool.setuptools.package-data]
"*" = ["openenv.yaml"]
